requirements.txt
dvc>=2.9.0
dvclive
dvc[s3]
dvc[gs]
evidently
fastapi
flask
flask_cors
flask_migrate
flask_sqlalchemy
matplotlib
mlem
numpy
pandas
pickle5
python-dotenv
sklearn
torch
torchvision
uvicorn
wheel
